cmd line utility to remove ghost information
